% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype-package.R
\docType{package}
\name{celltype-package}
\alias{celltype}
\alias{celltype-package}
\title{celltype: Prediction and analysis of cell types from single cell experiments}
\description{
Prediction and analysis of cell types from single cell experiments.
}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com}

}
\keyword{internal}
