% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{choose_celltype}
\alias{choose_celltype}
\alias{choose_celltype.matrix}
\title{choose_celltype}
\usage{
choose_celltype(x)

\method{choose_celltype}{matrix}(x)
}
\arguments{
\item{x}{a matrix of cell type predictions.}
}
\description{
Make decision about cell type identity based on prediction matrix.
}
