% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fix_immgen_celltype}
\alias{fix_immgen_celltype}
\alias{fix_immgen_celltype.SingleCellExperiment}
\alias{fix_immgen_celltype.data.frame}
\title{fix_immgen_celltype}
\usage{
fix_immgen_celltype(x)

\method{fix_immgen_celltype}{SingleCellExperiment}(x,
  colname = "celltype_immgen")

\method{fix_immgen_celltype}{data.frame}(x, colname = "celltype_immgen")
}
\arguments{
\item{x}{data.frame with cell type predictions.}

\item{colname}{name of column to fix.}
}
\description{
Fixes the cell type names in the immgen dataset by picking the upper level
cell type in the hierarchy.
}
