% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_celltype}
\alias{get_celltype}
\title{Returns a list of the celltypes available in the specified dictionary.}
\usage{
get_celltype(name, tissue = NULL)
}
\arguments{
\item{name}{name of dictionary.}

\item{tissue}{name of tissue.}
}
\description{
Returns a list of the celltypes available in the specified dictionary.
}
