% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_db}
\alias{get_db}
\title{get_db}
\usage{
get_db(name = "immgen", org = "mouse", tissue = NULL)
}
\arguments{
\item{name}{name of database.}

\item{org}{name of organism.}

\item{tissue}{name of tissue.}
}
\description{
get_db
}
