% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_tissue}
\alias{get_tissue}
\title{Returns a list of the tissues available in the specified dictionary.}
\usage{
get_tissue(name)
}
\arguments{
\item{name}{name of dictionary.}
}
\description{
Returns a list of the tissues available in the specified dictionary.
}
