% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simplify_immgen_celltype}
\alias{simplify_immgen_celltype}
\alias{simplify_immgen_celltype.SingleCellExperiment}
\alias{simplify_immgen_celltype.data.frame}
\title{simplify_immgen_celltype}
\usage{
simplify_immgen_celltype(x, colname = "celltype")

\method{simplify_immgen_celltype}{SingleCellExperiment}(x,
  colname = "celltype")

\method{simplify_immgen_celltype}{data.frame}(x, colname = "celltype")
}
\arguments{
\item{x}{data.frame with cell type predictions.}

\item{colname}{name of column to fix.}
}
\description{
Simplifies cell type names in the immgen dataset by picking the upper level
cell type in the hierarchy.
}
