% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_celltype}
\alias{predict_celltype}
\alias{predict_celltype.SingleCellExperiment}
\alias{predict_celltype.ExpressionSet}
\alias{predict_celltype.matrix}
\title{predict_celltyp}
\usage{
predict_celltype(x, org = "human", name = "immgen", db = NULL)

\method{predict_celltype}{SingleCellExperiment}(x, org = "human",
  name = "immgen", db = NULL, assay.name = "logcounts")

\method{predict_celltype}{ExpressionSet}(x, org = "human",
  name = "immgen", db = NULL)

\method{predict_celltype}{matrix}(x, db = NULL, name = "immgen",
  org = "human")
}
\arguments{
\item{x}{object with count matrix.}

\item{org}{target organism.}

\item{name}{name of cell type dictionary.}

\item{db}{a matrix cell type dictionary.}

\item{assay.name}{name of the assay to obtain expression matrix.}
}
\description{
predict_celltyp
}
