% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_celltype}
\alias{predict_celltype}
\alias{predict_celltype.SingleCellExperiment}
\alias{predict_celltype.ExpressionSet}
\alias{predict_celltype.matrix}
\title{predict_celltyp}
\usage{
predict_celltype(x, org = "mouse", name = "immgen", db = NULL, ...)

\method{predict_celltype}{SingleCellExperiment}(x, org = "mouse",
  name = "immgen", db = NULL, assay.name = "logcounts", ...)

\method{predict_celltype}{ExpressionSet}(x, org = "mouse",
  name = "immgen", db = NULL, ...)

\method{predict_celltype}{matrix}(x, org = "mouse", name = "immgen",
  db = NULL, ...)
}
\arguments{
\item{x}{object with count matrix.}

\item{org}{target organism.}

\item{name}{name of cell type dictionary.}

\item{db}{a matrix cell type dictionary.}

\item{...}{arguments passed down to methods.}

\item{assay.name}{name of the assay to obtain expression matrix.}
}
\description{
predict_celltyp
}
