% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_celltype}
\alias{predict_celltype}
\alias{predict_celltype.SingleCellExperiment}
\alias{predict_celltype.ExpressionSet}
\alias{predict_celltype.matrix}
\title{predict_celltyp}
\usage{
predict_celltype(x, db = NULL, name = "immgen", org = "human")

\method{predict_celltype}{SingleCellExperiment}(x, db = NULL,
  name = "immgen", org = "human", assay.name = "logcounts")

\method{predict_celltype}{ExpressionSet}(x, db = NULL, name = "immgen",
  org = "human")

\method{predict_celltype}{matrix}(x, db = NULL, name = "immgen",
  org = "human")
}
\description{
predict_celltyp
}
