% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simplify_immgen_celltype}
\alias{simplify_immgen_celltype}
\alias{simplify_immgen_celltype.character}
\title{simplify_immgen_celltype}
\usage{
simplify_immgen_celltype(x)

\method{simplify_immgen_celltype}{character}(x)
}
\arguments{
\item{x}{character vector with cell type names from Immgen.}
}
\description{
Simplifies cell type names in the immgen dataset by picking the upper level
cell type in the hierarchy.
}
